refactor(import): generalize import package across agents (Claude first)
Renames cmd/entire/cli/importclaude -> cmd/entire/cli/agentimport with an
Importer seam (Discover + SplitTurns) and a static set of supported agents,
so other agents can be added without touching the orchestration. Claude Code
is the only implementation today.

Also drops the enable-time import prompt; importing is a hidden command only
(entire import <agent>, discoverable via entire labs).

Entire-Checkpoint: e58fc114f6a9
